pyproject.toml
[tool.poetry]
name = "gpt-prompt-chainer"
version = "0.1.0"
description = ""
authors = ["Yauheni Bokach <dharmagetic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.30.5"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
watchdog = "^4.0.1"
pre-commit = "^3.7.1"
pytest = "^8.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
